refactor(ui): update test directory exports
