+++ packages/ui/src/test/index.ts
@@ -1,0 +1,1 @@
+export { default as render } from "./render";
